+++ GQNN/data/dataset.py
@@ -1,6 +1,3 @@
-# import os
-
-
 class Data_Read:
     """
     A utility class for reading, cleaning, manipulating, and scaling datasets.